Refactor for wso2-extension to Siddhi-io repo move
