+++ component/pom.xml
@@ -19,7 +19,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
-        <groupId>org.wso2.extension.siddhi.execution.regex</groupId>
+        <groupId>io.siddhi.extension.execution.regex</groupId>
         <artifactId>siddhi-execution-regex-parent</artifactId>
         <version>5.0.2-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
@@ -85,7 +85,7 @@
                         <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
                         <Bundle-Name>${project.artifactId}</Bundle-Name>
                         <Export-Package>
-                            org.wso2.extension.siddhi.execution.regex.*
+                            io.siddhi.extension.execution.regex.*
                         </Export-Package>
                         <Import-Package>
                             io.siddhi.core.*;version="${siddhi.version.range}",
