+++ component/src/main/java/io/siddhi/extension/execution/regex/FindFunctionExtension.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.extension.siddhi.execution.regex;
+package io.siddhi.extension.execution.regex;
 
 import io.siddhi.annotation.Example;
 import io.siddhi.annotation.Extension;
