+++ component/src/main/java/io/siddhi/extension/execution/regex/MatchesFunctionExtension.java
@@ -17,7 +17,7 @@
  */
 
 
-package org.wso2.extension.siddhi.execution.regex;
+package io.siddhi.extension.execution.regex;
 
 import io.siddhi.annotation.Example;
 import io.siddhi.annotation.Extension;
