+++ component/src/test/java/io/siddhi/extension/execution/regex/GroupFunctionExtensionTestCase.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.extension.siddhi.execution.regex;
+package io.siddhi.extension.execution.regex;
 
 import io.siddhi.core.SiddhiAppRuntime;
 import io.siddhi.core.SiddhiManager;
