+++ component/src/test/resources/testng.xml
@@ -3,12 +3,12 @@
 <suite name="RegexExpressionExtensionTestSuit" verbose="1">
     <test name="RegexExpressionExtensionTest">
         <classes>
-            <class name="org.wso2.extension.siddhi.execution.regex.RegExFunctionExtensionTestCase"/>
-            <class name="org.wso2.extension.siddhi.execution.regex.FindFunctionExtensionTestCase"/>
-            <class name="org.wso2.extension.siddhi.execution.regex.GroupFunctionExtensionTestCase"/>
-            <class name="org.wso2.extension.siddhi.execution.regex.LookingAtFunctionExtensionTestCase"/>
-            <class name="org.wso2.extension.siddhi.execution.regex.MatchesFunctionExtensionTestCase"/>
-            <class name="org.wso2.extension.siddhi.execution.regex.RegexPersistanceTestCase"/>
+            <class name="io.siddhi.extension.execution.regex.RegExFunctionExtensionTestCase"/>
+            <class name="io.siddhi.extension.execution.regex.FindFunctionExtensionTestCase"/>
+            <class name="io.siddhi.extension.execution.regex.GroupFunctionExtensionTestCase"/>
+            <class name="io.siddhi.extension.execution.regex.LookingAtFunctionExtensionTestCase"/>
+            <class name="io.siddhi.extension.execution.regex.MatchesFunctionExtensionTestCase"/>
+            <class name="io.siddhi.extension.execution.regex.RegexPersistanceTestCase"/>
         </classes>
     </test>
 </suite>
